requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
DateTime==5.4
Flask==3.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.5
numpy==1.26.4bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
DateTime==5.4
Flask==3.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
setuptools==69.2.0
Werkzeug==3.0.1
zope.interface==6.2

opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
setuptools==69.2.0
Werkzeug==3.0.1
zope.interface==6.2
